requirements.txt
requests
pyyaml
restview
doc8
jsonschema
